Refactor getReceipt
